Building AI Agents/2_autoRagPhiData/requirements.txt
phidata 
nest_asyncio 
beautifulsoup4 
openai 
duckduckgo-search 
sqlalchemy 
pgvector 
psycopg 
pypdf 
ollama